Breaking response to make it more versatile
